git-credential-manager: install successor tool
The Git Credential Manager for Mac and Linux tool currently installed with
this formula was archived ~3 years ago. It was superceded by Git
Credential Manager [1], which until now users have installed from a cask
in the microsoft/homebrew-git tap [2]. This change re-enables the formula
with the correct updates to allow x64/ARM64 users on macOS and Linux to
install Git Credential Manager from homebrew-core.

1: https://github.com/git-ecosystem/git-credential-manageri
2: https://github.com/microsoft/homebrew-git